+++ Gotcha/Views/LoginView.swift
@@ -14,7 +14,7 @@ struct LoginView: View {
 
     var body: some View {
         ZStack {
-            Color(red: 0.07, green: 0.07, blue: 0.10)
+            Theme.bg
                 .ignoresSafeArea()
 
             AnimatedBlobs()